backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
requests
joblib
lightgbm
pytest
pytest-asyncio
pytest-mock
httpx
pytest-cov
faker
python-dotenv
black
flake8
mypy
python-jose[cryptography]
passlib[bcrypt]
dependency-injector
tenacity
python-multipart
